pyproject.toml
[project]
name = "zorro-experiments"
version = "0.1.0"
description = "Common core foundation for ZORRO adaptation experiments"
requires-python = ">=3.9"
authors = [
    { name = "Research Automation" }
]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "scikit-learn>=1.2",
    "rich>=13.0",
    "tqdm>=4.0",
    "datasets>=2.12",
]

[build-system]
requires = ["setuptools", "wheel"]
